update example to test v7.9

+++ example/header.h
@@ -9,11 +9,11 @@ struct s_test {
 	int		val;
 };
 
-int	print_object(void *restrict obj_data, void *restrict any_data, size_t idx);
-int	match_slash(void *restrict obj_data, void *restrict any_data, size_t idx);
-int	match_obj1(void *restrict obj_data, void *restrict any_data, size_t idx);
-int	match_obj2(void *restrict obj_data, void *restrict any_data, size_t idx);
-int	match_obj2err(void *restrict obj_data, void *restrict any_data, size_t idx);
+ssize_t	print_object(void *restrict obj_data, void *restrict any_data, size_t idx);
+ssize_t	match_slash(void *restrict obj_data, void *restrict any_data, size_t idx);
+ssize_t	match_obj1(void *restrict obj_data, void *restrict any_data, size_t idx);
+ssize_t	match_obj2(void *restrict obj_data, void *restrict any_data, size_t idx);
+ssize_t	match_obj2err(void *restrict obj_data, void *restrict any_data, size_t idx);
 
 void	free_obj(void *restrict data);
